Move string literal initialisation into separate file
